refactor: rename _cardfit to _fit — it fits every widget family now
The module stopped being card-specific when the gauge/progress
estimator fork was retired; the name lied to the next reader. Mechanical
rename, imports and docs updated, no behavior change.
