+++ custom_components/geekmagic/widgets/entity.py
@@ -11,13 +11,13 @@ from ..const import (
 from ..htmldoc import css_rgb
 from ._bands import plan_bands
 from ._card import card_html
-from ._cardfit import (
+from ._cellkit import cell_box, label_px
+from ._fit import (
     HERO_SHARE_SOLO,
     HERO_SHARE_STACKED,
     fit_hero,
     hero_block,
 )
-from ._cellkit import cell_box, label_px
 from .base import Widget, WidgetConfig
 from .helpers import get_binary_sensor_icon, translate_binary_state
