+++ custom_components/geekmagic/widgets/text.py
@@ -8,13 +8,13 @@ from ..const import PLACEHOLDER_VALUE
 from ..htmldoc import css_rgb
 from ._bands import plan_bands
 from ._card import card_html
-from ._cardfit import (
+from ._cellkit import cell_box, label_px
+from ._fit import (
     HERO_SHARE_SOLO,
     HERO_SHARE_STACKED,
     fit_hero,
     hero_block,
 )
-from ._cellkit import cell_box, label_px
 from .base import Widget, WidgetConfig
 
 if TYPE_CHECKING:
