+++ custom_components/geekmagic/widgets/weather.py
@@ -7,14 +7,14 @@ from html import escape
 from typing import TYPE_CHECKING, Any, ClassVar
 
 from ..htmldoc import mdi_span
-from ._cardfit import (
+from ._cellkit import caption_visible, cell_box, chip_band_px, label_px
+from ._fit import (
     HERO_SHARE_SOLO,
     HERO_SHARE_STACKED,
     fit_caption_sized,
     fit_hero,
     hero_block,
 )
-from ._cellkit import caption_visible, cell_box, chip_band_px, label_px
 from .base import Widget, WidgetConfig
 from .state import DataNeeds
